# block refresh for invisible layers