--- a/napari/_vispy/vispy_points_layer.py
+++ b/napari/_vispy/vispy_points_layer.py
@@ -1,5 +1,6 @@
 import numpy as np
-from vispy.scene.visuals import Line, Compound, Markers
+from vispy.scene.visuals import Line, Compound
+from .markers import Markers
 from vispy.visuals.transforms import ChainTransform
 from .vispy_base_layer import VispyBaseLayer
 
@@ -37,7 +38,6 @@ class VispyPointsLayer(VispyBaseLayer):
             self.node = Compound([Markers(), Markers(), Line()])
         else:
             self.node = Markers()
-
         self.node.parent = parent
         self._reset_base()
 
@@ -119,8 +119,11 @@ class VispyPointsLayer(VispyBaseLayer):
             scaling=True,
         )
 
-        if 0 in self.layer._highlight_box.shape:
-            pos = np.zeros((1, 2))
+        if (
+            self.layer._highlight_box is None
+            or 0 in self.layer._highlight_box.shape
+        ):
+            pos = np.zeros((1, self.layer.dims.ndisplay))
             width = 0
         else:
             pos = self.layer._highlight_box