--- a/napari/layers/shapes/shapes.py
+++ b/napari/layers/shapes/shapes.py
@@ -567,7 +567,7 @@ class Shapes(Layer):
         self.events.mode(mode=mode)
         if not (mode in draw_modes and old_mode in draw_modes):
             self._finish_drawing()
-        self.set_view_slice()
+        self.refresh()
 
     def _set_editable(self, editable=None):
         """Set editable mode based on layer properties."""
@@ -1125,7 +1125,7 @@ class Shapes(Layer):
         new_z_index = max(self._data_view._z_index) + 1
         for index in self.selected_data:
             self._data_view.update_z_index(index, new_z_index)
-        self.set_view_slice()
+        self.refresh()
 
     def move_to_back(self):
         """Moves selected objects to be displayed behind all others."""
@@ -1134,7 +1134,7 @@ class Shapes(Layer):
         new_z_index = min(self._data_view._z_index) - 1
         for index in self.selected_data:
             self._data_view.update_z_index(index, new_z_index)
-        self.set_view_slice()
+        self.refresh()
 
     def _copy_data(self):
         """Copy selected shapes to clipboard."""
@@ -1198,7 +1198,7 @@ class Shapes(Layer):
                     for index in self.selected_data:
                         self._data_view.shift(index, shift)
                     self._selected_box = self._selected_box + shift
-                    self.set_view_slice()
+                    self.refresh()
                 elif vertex < Box.LEN:
                     # Corner / edge vertex is being dragged so resize object
                     box = self._selected_box
@@ -1276,7 +1276,7 @@ class Shapes(Layer):
                         self._transform_box(
                             transform, center=self._fixed_vertex
                         )
-                    self.set_view_slice()
+                    self.refresh()
                 elif vertex == 8:
                     # Rotation handle is being dragged so rotate object
                     handle = self._selected_box[Box.HANDLE]
@@ -1308,7 +1308,7 @@ class Shapes(Layer):
                             index, angle, center=self._fixed_vertex
                         )
                     self._rotate_box(angle, center=self._fixed_vertex)
-                    self.set_view_slice()
+                    self.refresh()
             else:
                 self._is_selecting = True
                 if self._drag_start is None:
@@ -1338,7 +1338,7 @@ class Shapes(Layer):
                         )
                         shapes = self.selected_data
                         self._selected_box = self.interaction_box(shapes)
-                        self.set_view_slice()
+                        self.refresh()
             else:
                 self._is_selecting = True
                 if self._drag_start is None:
@@ -1482,7 +1482,7 @@ class Shapes(Layer):
             self._value = (self.selected_data[0], 4)
             self._moving_value = copy(self._value)
             self._is_creating = True
-            self.set_view_slice()
+            self.refresh()
         elif self._mode in [Mode.ADD_PATH, Mode.ADD_POLYGON]:
             if self._is_creating is False:
                 # Start drawing a path
@@ -1581,7 +1581,7 @@ class Shapes(Layer):
                 data_full = self.expand_shape(vertices)
                 self._data_view.edit(index, data_full, new_type=new_type)
                 self._selected_box = self.interaction_box(self.selected_data)
-            self.set_view_slice()
+            self.refresh()
         elif self._mode == Mode.VERTEX_REMOVE:
             if self._value[1] is not None:
                 # have clicked on a current vertex so remove
@@ -1624,7 +1624,7 @@ class Shapes(Layer):
                         )
                         shapes = self.selected_data
                         self._selected_box = self.interaction_box(shapes)
-                self.set_view_slice()
+                self.refresh()
         else:
             raise ValueError("Mode not recongnized")