--- a/napari/layers/surface/surface.py
+++ b/napari/layers/surface/surface.py
@@ -317,7 +317,7 @@ class Surface(Layer):
 
         return xml_list
 
-    def get_value(self):
+    def _get_value(self):
         """Returns coordinates, values, and a string for a given mouse position
         and set of indices.