… benchmarks/benchmark_image_layer.py
… benchmarks/benchmark_labels_layer.py
… benchmarks/benchmark_points_layer.py
… benchmarks/benchmark_qt_viewer_image.py
… benchmarks/benchmark_qt_viewer_labels.py
… benchmarks/benchmark_shapes_layer.py
… benchmarks/benchmark_surface_layer.py
… benchmarks/benchmark_vectors_layer.py
… napari/layers/image/tests/test_big_image_timing.py
… napari/layers/surface/tests/test_surface.py
… napari/layers/vectors/vectors.py
… napari/tests/test_viewer.py
… napari/tests/util.py