# Description
This PR will close #578 and block refreshes for invisible layers. I'd still like to add some tests, add some benchmarks and do some profiling before we merge, but I think it's there from a functionality standpoint and it is actually a pretty minimal change. Note that for nD data the thumbnail won't update for an invisible layer as the dims slider moves, but that seems reasonable. I might need to do more blocking on the vispy side of things, but I'm not sure.

Any feedback from @jni, @eburling (who was running into this problem), or others at this point would be welcome. 

## Type of change
<!-- Please delete options that are not relevant. -->
- [x] Bug-fix (non-breaking change which fixes an issue)