@sofroniewn 2019-12-12
Here are results for the updated benchmarks

Baseline results when the layer is visible (by default)
```
(base) czipa1osx041:napari nicholassofroniew$ asv run -E existing -b benchmark_qt_viewer_image.QtViewerSingleImageSuite
· Discovering benchmarks
· Running 7 total benchmarks (1 commits * 1 environments * 7 benchmarks)
[  0.00%] ·· Benchmarking existing-py_anaconda3_bin_python
[  7.14%] ··· Running (benchmark_qt_viewer_image.QtViewerSingleImageSuite.time_get_value--).......
[ 57.14%] ··· benchmark_qt_viewer_image.QtViewerSingleImageSuite.time_get_value                                                                                                                  16.1±0.09μs
[ 64.29%] ··· benchmark_qt_viewer_image.QtViewerSingleImageSuite.time_ndisplay                                                                                                                      872±10ns
[ 71.43%] ··· benchmark_qt_viewer_image.QtViewerSingleImageSuite.time_refresh                                                                                                                    1.05±0.01ms
[ 78.57%] ··· benchmark_qt_viewer_image.QtViewerSingleImageSuite.time_set_data                                                                                                                   1.31±0.01ms
[ 85.71%] ··· benchmark_qt_viewer_image.QtViewerSingleImageSuite.time_set_view_slice                                                                                                              54.9±0.8μs
[ 92.86%] ··· benchmark_qt_viewer_image.QtViewerSingleImageSuite.time_update_thumbnail                                                                                                               749±5μs
[100.00%] ··· benchmark_qt_viewer_image.QtViewerSingleImageSuite.time_zoom                                                                                                                       1.46±0.01ms
```

New results when the layer is not visible
```
(base) czipa1osx041:napari nicholassofroniew$ asv run -E existing -b benchmark_qt_viewer_image.QtViewerSingleInvisbleImageSuite
· Discovering benchmarks
· Running 7 total benchmarks (1 commits * 1 environments * 7 benchmarks)
[  0.00%] ·· Benchmarking existing-py_anaconda3_bin_python
[  7.14%] ··· Running (benchmark_qt_viewer_image.QtViewerSingleInvisbleImageSuite.time_get_value--)......
[ 50.00%] ··· Running (benchmark_qt_viewer_image.QtViewerSingleInvisbleImageSuite.time_zoom--).
[ 57.14%] ··· benchmark_qt_viewer_image.QtViewerSingleInvisbleImageSuite.time_get_value                                                                                                          2.08±0.07μs
[ 64.29%] ··· benchmark_qt_viewer_image.QtViewerSingleInvisbleImageSuite.time_ndisplay                                                                                                              507±20ns
[ 71.43%] ··· benchmark_qt_viewer_image.QtViewerSingleInvisbleImageSuite.time_refresh                                                                                                             2.15±0.1μs
[ 78.57%] ··· benchmark_qt_viewer_image.QtViewerSingleInvisbleImageSuite.time_set_data                                                                                                              222±10μs
[ 85.71%] ··· benchmark_qt_viewer_image.QtViewerSingleInvisbleImageSuite.time_set_view_slice                                                                                                        56.0±2μs
[ 92.86%] ··· benchmark_qt_viewer_image.QtViewerSingleInvisbleImageSuite.time_update_thumbnail                                                                                                      790±40μs
[100.00%] ··· benchmark_qt_viewer_image.QtViewerSingleInvisbleImageSuite.time_zoom                                                                                                                1.58±0.1ms
```

Note that the time for refresh is now much smaller when the layer is not visible. Also note though that the zoom time is the same. This might be because the zoom time isn't capturing work done on the gpu. If you look at the benchmarks below the zoom time also remains constant as size of data increases, which seems unlikely

```
(base) czipa1osx041:napari nicholassofroniew$ asv run -E existing -b benchmark_qt_viewer_image.QtViewerImageSuite
· Discovering benchmarks
· Running 5 total benchmarks (1 commits * 1 environments * 5 benchmarks)
[  0.00%] ·· Benchmarking existing-py_anaconda3_bin_python
[ 10.00%] ··· Running (benchmark_qt_viewer_image.QtViewerImageSuite.time_get_value--).
[ 20.00%] ··· Running (benchmark_qt_viewer_image.QtViewerImageSuite.time_refresh--).
[ 30.00%] ··· Running (benchmark_qt_viewer_image.QtViewerImageSuite.time_set_view_slice--).
[ 40.00%] ··· Running (benchmark_qt_viewer_image.QtViewerImageSuite.time_update_thumbnail--).
[ 50.00%] ··· Running (benchmark_qt_viewer_image.QtViewerImageSuite.time_zoom--).
[ 60.00%] ··· benchmark_qt_viewer_image.QtViewerImageSuite.time_get_value                                                                                                                                 ok
[ 60.00%] ··· ======== ============
               param1
              -------- ------------
                 16     16.3±0.2μs
                 32     16.3±0.4μs
                 64     16.0±0.8μs
                128     16.2±0.2μs
                256     16.0±0.7μs
                512     16.5±0.3μs
                1024    16.0±0.2μs
                2048    15.9±0.1μs
                4096    16.1±0.2μs
              ======== ============

[ 70.00%] ··· benchmark_qt_viewer_image.QtViewerImageSuite.time_refresh                                                                                                                                   ok
[ 70.00%] ··· ======== =============
               param1
              -------- -------------
                 16       927±20μs
                 32       925±6μs
                 64       943±40μs
                128       960±70μs
                256       985±9μs
                512     1.18±0.02ms
                1024     1.98±0.2ms
                2048     8.98±0.3ms
                4096      45.3±1ms
              ======== =============

[ 80.00%] ··· benchmark_qt_viewer_image.QtViewerImageSuite.time_set_view_slice                                                                                                                            ok
[ 80.00%] ··· ======== ============
               param1
              -------- ------------
                 16      15.6±1μs
                 32      15.3±3μs
                 64     15.4±0.2μs
                128     15.1±0.2μs
                256     15.7±0.9μs
                512     15.4±0.7μs
                1024    15.2±0.3μs
                2048    15.3±0.3μs
                4096    15.1±0.1μs
              ======== ============

[ 90.00%] ··· benchmark_qt_viewer_image.QtViewerImageSuite.time_update_thumbnail                                                                                                                          ok
[ 90.00%] ··· ======== ==========
               param1
              -------- ----------
                 16     720±4μs
                 32     746±4μs
                 64     746±7μs
                128     755±7μs
                256     746±4μs
                512     754±30μs
                1024    841±90μs
                2048    807±50μs
                4096    771±10μs
              ======== ==========

[100.00%] ··· benchmark_qt_viewer_image.QtViewerImageSuite.time_zoom                                                                                                                                      ok
[100.00%] ··· ======== =============
               param1
              -------- -------------
                 16     1.43±0.02ms
                 32     1.45±0.09ms
                 64     1.42±0.02ms
                128     1.44±0.02ms
                256     1.41±0.01ms
                512     1.43±0.03ms
                1024    1.41±0.01ms
                2048    1.42±0.01ms
                4096    1.44±0.01ms
              ======== =============
```